[#3] Implement auth guards using Next.JS built-in middleware
