--- nextjs/middleware.ts
@@ -1,15 +1,0 @@
-import { getIronSession } from 'iron-session/edge';
-import type { NextRequest } from 'next/server';
-import { NextResponse } from 'next/server';
-
-import { sessionOptions } from './src/config/session';
-
-export const middleware = async (req: NextRequest) => {
-  const res = NextResponse.next();
-  const session = await getIronSession(req, res, sessionOptions);
-  const { user } = session;
-
-  console.log('nextjs middleware', user);
-
-  return res;
-};
