+++ nextjs/src/lib/middleware/passport.middleware.ts
@@ -1,17 +1,9 @@
 import passport from 'passport';
 import { Strategy as GoogleStrategy } from 'passport-google-oauth20';
 
-import { SerializedUser } from '../../models/user.model';
 import * as UserRepository from '../../repositories/user.repository';
 import AuthError from '../../services/auth/error';
 import AuthGoogleService from '../../services/auth/google.service';
-
-declare global {
-  // eslint-disable-next-line @typescript-eslint/no-namespace
-  namespace Express {
-    type User = SerializedUser;
-  }
-}
 
 passport.serializeUser(({ id }, callback) => {
   callback(null, id);
