+++ nextjs/src/pages/auth/sign-in.tsx
@@ -1,0 +1,19 @@
+import type { NextPage } from 'next';
+import Head from 'next/head';
+import Link from 'next/link';
+
+const AuthSignInPage: NextPage = () => {
+  return (
+    <>
+      <Head>
+        <title>Sign in | NextNewsletter 🚀</title>
+      </Head>
+
+      <Link href="../api/auth/sign-in">
+        <button className="btn">Login</button>
+      </Link>
+    </>
+  );
+};
+
+export default AuthSignInPage;
